Migrate AdminConsoleLandingPageTest to the new framework
Part of #34494
